fix: move data collection methods to stonecrop class
